+++ src/path_aggregation.hpp
@@ -43,8 +43,8 @@ public:
 	void enqueue(
 		const feature_type *left,
 		const feature_type *right,
-		size_t width,
-		size_t height,
+		int width,
+		int height,
 		unsigned int p1,
 		unsigned int p2,
 		cudaStream_t stream);
