+++ src/vertical_path_aggregation.cu
@@ -29,8 +29,8 @@ __global__ void aggregate_vertical_path_kernel(
 	uint8_t *dest,
 	const feature_type *left,
 	const feature_type *right,
-	unsigned int width,
-	unsigned int height,
+	int width,
+	int height,
 	unsigned int p1,
 	unsigned int p2)
 {
@@ -116,8 +116,8 @@ void enqueue_aggregate_up2down_path(
 	cost_type *dest,
 	const feature_type *left,
 	const feature_type *right,
-	size_t width,
-	size_t height,
+	int width,
+	int height,
 	unsigned int p1,
 	unsigned int p2,
 	cudaStream_t stream)
@@ -125,8 +125,8 @@ void enqueue_aggregate_up2down_path(
 	static const unsigned int SUBGROUP_SIZE = MAX_DISPARITY / DP_BLOCK_SIZE;
 	static const unsigned int PATHS_PER_BLOCK = BLOCK_SIZE / SUBGROUP_SIZE;
 
-	const size_t gdim = (width + PATHS_PER_BLOCK - 1) / PATHS_PER_BLOCK;
-	const size_t bdim = BLOCK_SIZE;
+	const int gdim = (width + PATHS_PER_BLOCK - 1) / PATHS_PER_BLOCK;
+	const int bdim = BLOCK_SIZE;
 	aggregate_vertical_path_kernel<1, MAX_DISPARITY><<<gdim, bdim, 0, stream>>>(
 		dest, left, right, width, height, p1, p2);
 }
@@ -136,8 +136,8 @@ void enqueue_aggregate_down2up_path(
 	cost_type *dest,
 	const feature_type *left,
 	const feature_type *right,
-	size_t width,
-	size_t height,
+	int width,
+	int height,
 	unsigned int p1,
 	unsigned int p2,
 	cudaStream_t stream)
@@ -145,8 +145,8 @@ void enqueue_aggregate_down2up_path(
 	static const unsigned int SUBGROUP_SIZE = MAX_DISPARITY / DP_BLOCK_SIZE;
 	static const unsigned int PATHS_PER_BLOCK = BLOCK_SIZE / SUBGROUP_SIZE;
 
-	const size_t gdim = (width + PATHS_PER_BLOCK - 1) / PATHS_PER_BLOCK;
-	const size_t bdim = BLOCK_SIZE;
+	const int gdim = (width + PATHS_PER_BLOCK - 1) / PATHS_PER_BLOCK;
+	const int bdim = BLOCK_SIZE;
 	aggregate_vertical_path_kernel<-1, MAX_DISPARITY><<<gdim, bdim, 0, stream>>>(
 		dest, left, right, width, height, p1, p2);
 }
@@ -156,8 +156,8 @@ template void enqueue_aggregate_up2down_path<64u>(
 	cost_type *dest,
 	const feature_type *left,
 	const feature_type *right,
-	size_t width,
-	size_t height,
+	int width,
+	int height,
 	unsigned int p1,
 	unsigned int p2,
 	cudaStream_t stream);
@@ -166,8 +166,8 @@ template void enqueue_aggregate_up2down_path<128u>(
 	cost_type *dest,
 	const feature_type *left,
 	const feature_type *right,
-	size_t width,
-	size_t height,
+	int width,
+	int height,
 	unsigned int p1,
 	unsigned int p2,
 	cudaStream_t stream);
@@ -176,8 +176,8 @@ template void enqueue_aggregate_down2up_path<64u>(
 	cost_type *dest,
 	const feature_type *left,
 	const feature_type *right,
-	size_t width,
-	size_t height,
+	int width,
+	int height,
 	unsigned int p1,
 	unsigned int p2,
 	cudaStream_t stream);
@@ -186,8 +186,8 @@ template void enqueue_aggregate_down2up_path<128u>(
 	cost_type *dest,
 	const feature_type *left,
 	const feature_type *right,
-	size_t width,
-	size_t height,
+	int width,
+	int height,
 	unsigned int p1,
 	unsigned int p2,
 	cudaStream_t stream);
